Changed import of Aer and IBMQ backend
Aer is now imported via qiskit_aer and the IBMQ backend has been changed to IBMProvider and can be imported from qiskit_ibm_provider. Some methods have been changed or deleted in the IBMProvider module, which were previously a part of the IBMQ object. Hence, some minor changes in the code had to be done.